requirements/base.txt
Django==1.10.3
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.3
django_cors_headers==1.3.1
drf_extensions==0.3.1
psycopg2==2.6.2
six==1.10.0
uwsgi==2.0.14
pyyaml==3.12
pysqlite===2.8.3
ipdb==0.10.1
celery==4.0.0
librabbitmq==1.6.1
kombu==4.0.0
django-celery-results==1.0.1
django-extensions==1.7.4
django-grappelli==2.9.1
path.py==9.0
pickleshare==0.7.4
Pillow==3.4.2
python-dateutil==2.6.0
python-memcached==1.58
requests==2.12.1
simplegeneric==0.8.1
traitlets==4.3.1
